requirements_week1.txt
flask
opensearch-py
ipython
python-dotenv